Remove dot net 2.0 since it's out of support and build succesfully.
