--- YamlDotNet/Portability/net20/include/HashSet.cs
@@ -1,79 +1,0 @@
-﻿// This file is part of YamlDotNet - A .NET library for YAML.
-// Copyright (c) Antoine Aubry and contributors
-//
-// Permission is hereby granted, free of charge, to any person obtaining a copy of
-// this software and associated documentation files (the "Software"), to deal in
-// the Software without restriction, including without limitation the rights to
-// use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies
-// of the Software, and to permit persons to whom the Software is furnished to do
-// so, subject to the following conditions:
-//
-// The above copyright notice and this permission notice shall be included in all
-// copies or substantial portions of the Software.
-//
-// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-// SOFTWARE.
-
-namespace System.Collections.Generic
-{
-    internal class HashSet<T> : IEnumerable<T>
-    {
-        private readonly Dictionary<T, object?> items;
-
-        public HashSet()
-        {
-            items = new Dictionary<T, object?>();
-        }
-
-        public HashSet(IEqualityComparer<T> comparer)
-        {
-            items = new Dictionary<T, object?>(comparer);
-        }
-
-        public HashSet(IEnumerable<T> collection) : this()
-        {
-            foreach (var item in collection)
-            {
-                Add(item);
-            }
-        }
-
-        public bool Add(T value)
-        {
-            if (Contains(value))
-            {
-                return false;
-            }
-            else
-            {
-                items.Add(value, null);
-                return true;
-            }
-        }
-
-        public bool Contains(T value)
-        {
-            return items.ContainsKey(value);
-        }
-
-        public void Clear()
-        {
-            items.Clear();
-        }
-
-        public IEnumerator<T> GetEnumerator()
-        {
-            return items.Keys.GetEnumerator();
-        }
-
-        IEnumerator IEnumerable.GetEnumerator()
-        {
-            return GetEnumerator();
-        }
-    }
-}
